requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
jwt==1.3.1
passlib==1.7.4
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing==3.7.4.3
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websockets==12.0
